requirements.txt
# Web Framework
Django>=5.0,<6.0

# REST API Framework
djangorestframework>=3.15.1,<4.0

# API Documentation
drf-yasg>=1.21.7

# OpenAI integration
openai>=1.12.0

# HTTP Server for Production Deployment
gunicorn>=21.2.0

# Environment Variables Management
python-dotenv>=1.0.1

# CORS handling
django-cors-headers>=4.3.1

# Testing
pytest>=8.0.1
pytest-django>=4.8.0
coverage>=7.4.1
factory-boy>=3.3.0
Faker>=22.6.0

# Progress bars
tqdm>=4.66.1

# Rate Limiting
django-ratelimit>=4.1.0

# Security
django-secure>=1.0.2
django-csp>=3.7

# Azure-specific packages
whitenoise>=6.6.0